layout: fix intrinsic size parameters from per axis to per frag
Intrinsic sizes are defined per element, not per axis.
That is, we can't render one axis with an intrinsic parameter and the other axis
with a different parameter.

That said, when refering to:
- sizing the flex container under a cross-axis min-content constraint
- The min/max-content size in the block/inline axis.
It is meant the whole element layout-ed into a min/max constraint and the axis
computed for this layout-ed element.