vaev-layout: fix table layer interaction
Before: table layers defined by rows/cols/groups were treated as boxes and
background interaction was in function of overlapping boxes.
This was wrong though because the background color of rows/cols/groups should
be only for the cells, not the whole rectangle defined by the cells (the said
boxes).

Now: background defined by rows/cols/groups are added to the cell style, what
will prevent the background printing on wrong places and the override of calling
layout for rows/cols/groups (what was also 'incorrect').

Note that, in borders collapse, rows/cols/groups can have borders and the same
strategy will need to be applied.