karm-print: parametrize stock paper size with scale
The Print::PaperStock _paper attribute from Karm::Scene::Page is sized taking
the scale parameter into account; however, this should not be the size of the
paper when printing, that is, we should use the original paper size, not the
scaled one.

This commit couples the scaling logic to the printing classes, allowing the
_paper attribute from Karm::Scene::Page to be sized back to its original size
and thus the correct dimenions of the printing paper to be used.