vaev-layout: Fix to use floor in flex-shrink instead of scaledShrinkFactor and floor other cases.
Related to: https://www.w3.org/TR/css-flexbox-1/#intrinsic-main-sizes

Previous implementation mis interpreated what specs meant when computing flex
fractions. That was fixed.
However, at step 3 of this same algo in the specs, we are supposed to perform
a division with a possible 0 value. Thus, the same flooring was done altho this
was not reiterated in the specs.